Minimum Number of Steps to Make Two Strings Anagram
This Java code defines a class named StringsAnagram with a method minSteps that calculates the minimum number of steps required to make two input strings, s and t, anagrams of each other. The method uses a HashMap to keep track of the frequency of characters in the second string t. It then iterates through the characters of the first string s and adjusts the frequency counts in the HashMap accordingly. The final result is the count of characters in s that are not present in t or have a different frequency. The returned value represents the minimum steps needed to make the two strings anagrams.