Climbing Stairs
Intuition
The problem can be solved using dynamic programming, where the key idea is to break down the problem into simpler subproblems and build the solution from the bottom up.

Approach
The climbStairs function uses an array ways to store the number of ways to climb each step. It starts by initializing the base cases (ways[0] and ways[1]) and iteratively calculates the number of ways for each step up to n by summing the values for the previous two steps.